Terms in teacher dashboard: fix printing & cookie expiration
Printing was failing in its JavaScript call because the iframe was loaded cross-domain.  Used the /dashboardapi entrypoint to load terms from the dashboard but under the code.org domain.

The cookie was being written as a session cookie because the window already had a setCookie function, and wasn't expiring at midnight.  Now fixed.