Allow tab-separated student first names
Because the full name splitter normalizes all whitespace to spaces, we
were unable to detect the uniqueness of first names that contained
non-space whitespace (because we were entirely unable to find the name
in our trie). By normalizing whitespace before we insert the names into
the trie, we're able to enforce consistency.