Fix some bad assumptions in playlab blocks
Namely, the assumption that our definition of skin will always be
fully-formed, with all the fields of all the types that we expect. This
assumption is valid in all of our Code Studio environments, but fails in
levelbuilder, where we initialize simple readonly blockspaces to display
block previews.

This assumption has been failing for a while, but for most of that time
this code was being loaded async, so when it broke it didn't break any
of the other javascript on the page. A recent change altered that,
revealing this problem.