pegasus: Move active drupal files into regular code.org site
We are logging to honeybadger as legacy Drupal-era assets are still being requested.  This moves a sampling of what seem to be the more popular files over into the code.org Pegasus site.

Unfortunately we don't know who's requesting them because CloudFront is stripping the request.  There's a valid question of whether we should just delete these files and return 404s instead.