Accomodate missing or malformed Callout I18n data
Specifically, instances where we expect an entry and there is none or
there is an entry but it is not (as expected) a hash. The latter can
happen when no subentries have yet been translated in crowdin; in that
case, because we have no way of specifying what form of data we want,
the YAML entry will default to being parsed as an empty Array rather
than an empty Hash.

Being in either situation simply means we have no translations for what
we want, so we default back to English.