Add context help
Rename the "Setup Status" section to "Setup Check."

Give each setup check step a help block that is displayed if that step fails. The idea is to give users a simpler debugging process by providing help related to the specific failure they're experiencing.

Moves the Setup Check above the general install instructions, since it's largely a replacement for them at this point and is usually the first thing you want to see when loading this page.