User type cookie: more improvements
Previously, the cookie deletion wasn't working.  After a fair investigation, have not managed to find a function to delete a cookie, but found success by setting the cookie to blank with an out-of-date expiration.

The cookie is now named _user_type, with a suffix for non-production environments, e.g. _user_type_development.

Like the _learn_session, the cookie now an HTTP-only cookie with a domain of .code.org