Print warning in Slack if rebase adds a stash
We had issues on Monday, April 3 2017 with content being stashed but not restored, in what appears to be a side effect of the `git pull --rebase --autostash` operation.  It looks like this is not a hard failure state but here I'd like to start logging warnings to Slack when this happens so we can get a feel for how much of a problem it is.