Volunteer: unsubscribe now detects non-18+ users and confirms age
Change https://github.com/code-dot-org/code-dot-org/pull/13643 added support for a new field that declares that the volunteer is 18+.  This caused an issue for the user experience in which a volunteer would attempt to unsubscribe at /volunteer/engineer/edit/[secret].  Attempting to unsubscribe would give an error, internally this was because the form wanted to know that the user was 18+.

This solution detects users who have not yet declared themselves as 18+ when showing the unsubscribe UI, and shows a checkbox for them.  If they attempt to submit the unsubscribe form without this checked, they will get a specific error message.

The unsubscribe form itself continue to not require an unsubscribe radio button to be checked, at least on the server-side processing of the form, and if no item is selected, then we write null to the database for their unsubscribe choice.  This has an odd side effect of allowing users to submit their 18+ status via the unbsubscribe form without actually unsubscribing, but we've deemed this okay.

The regular error message for this form now has a clickable link to open a support ticket.