Remove mkdirp dependency
We use this in exactly one place, in our Gruntfile.  We also do the exact same thing with `childProcess.execSync` in at least one other place in the Gruntfile.  Lots of our dependencies depend on this in turn, but it's such a small thing there's no need to depend on it directly.