Remove stage summary from script summary on edit page
A bit of poking around has revealed that, on environments without caching like
levelbuilder, building the stage summary for a script is ridiculously expensive,
and is very likely to blame for the levelbuilder timeouts we've been seeing
recently on larger scripts.

Since the stage summary is only used for the beta script editor which has been
broken for a while now, disable the stage summary so levelbuilders can be
unblocked in the short term, pending a longer-term investigation into the
current state of the beta editor.