Hamburger: proper ellipses and no more bump when sliding expander
This fixes the ellipses for both full width and narrow hamburger, so that they are shown for regular items, the expander item, and the subitems inside and expanded group.

It also removes the margin-top from the top item in the expanded group which fixes the "bump" in the slideToggle animation.