Header: increase availability of Sign In & hamburger
Until now, pages that used the sessions controller hid both the Sign In button and the hamburger.  In particular, this affected /users/sign_in and /oauth_sign_out/*.

With this change, the hamburger is now available on both pages, and the Sign In button is available on /users/sign_in.