Remove --out option from runner.rb
This option was originally added by @balderdash in https://github.com/code-dot-org/code-dot-org/pull/8789.  It's not used by any of our scripts, and after talking it over with him we're not sure anyone uses it manually, either.  It also doesn't work properly if you're running multiple features.  Removing it here to reduce moving parts in the UI test runner.