Support AppSpecificFailError in new Flappy levels
Formerly, exactly _one_ flappy level had an "appSpecificFailError" field, and we
were examining the level ID to determine if we wanted to consider it. Update the
logic to instead support this feature for any level that defines that field, and
add levelbuilder support for defining and editing.

Also ensure that failure/success conditions don't overwrite other goal
definitions.