Home: updated Leave Section now works
The Leave Section column is now shown for non-teacher-managed accounts.  Leaving a section is done with an ajax post and also just refreshes the entire /home page to get a fresh rendering of sections.  (The existing model API returns a redirect so far, which is no longer so useful for an ajax post.)  Only teachers will have links to sections in teacher-dashboard.