Home: fix course listing properly
This is the proper fix for the problem that https://github.com/code-dot-org/code-dot-org/pull/16587 attempted to fix.

The problem was that we were seeing the course featured as the student's primary course repeated in the list of recent courses below.

The real cause of the problem was the assumption that this primary course was the first of the recent courses.  The attempted fix concealed the problem by hiding another of the recent courses, but this didn't always work, and left the first of the recent courses never showing.

The proper fix is to go back to the dashboard code which removes the primary course from the recent courses, but actually removing the matching course, not just the first one in the list.