Make "Print certificates" button use absolute URL
This is a necessary prerequisite to moving the Section Table (which includes the Print Certificates button) over the dashboard.  Extracting this work from https://github.com/code-dot-org/code-dot-org/pull/16644 because it's valid in isolation.

Introduces a redux `urlHelpers` module that can currently be used to construct pegasus URLs, may be extended to help construct dashboard URLs in the future.  We thought this'd be cleaner than adding another value to pageConstants, which is more of a puzzle page concern anyway.

pair: @epeach