Courses: refactor the course blocks and make some updates
The course blocks remain as legacy HAML but are now managed by CourseBlocks, which provides several React components that provide common groupings of these blocks on /courses.

This also updates the view for students and logged-out users in English to see the newly-released CSF courses: Courses A-F, Express, and Pre-Express.