Courses: improve test for filtering primary script out of recent scripts
The unit test for ensuring that we can optionally filter the primary script out of the list of recent courses & scripts was failing 5-10% of the time, though notably all the time when debugging in RubyMine.  However, it was actually behaving differently to expected in the success case.

The root issue appears to have been non-deterministic setting of the primary script when using assign_script.  The solution is to more explicitly set started_at times for the user_scripts in question so that the primary_script retrieval becomes deterministic.