Header: The Pegasus header's user menu & hamburger no longer wrap
With a particularly long set of options in the Pegasus header on a responsive page, there was a chance that the user menu and hamburger would wrap into the whitespace below, rendering them invisible, at certain browser widths.

The same thing was occurring on the teacher-dashboard, since he header has remained responsive even as those pages themselves are not.

This fix brings the Pegasus functionality closer to the dashboard, by not allowing these right-hand components to wrap.  It does produce some overlapped rendering, but going forward we can tune the content of the header and the responsive rules to prevent that.