Perform one less navigation in several UI tests
Removes an unneeded navigation step from the `generate_teacher_student` helper that's used in many of our UI tests.  In local testing this saved about 15 seconds in one of our tests.

This helper gets used 23 times in our UI/eyes tests, so I'm hoping for a reduction of UI test runtime of nearly 6 minutes if tests were run in one thread (probably much less when we run in paralle, but still helpful).