Advance collector even on non-optimal completion
Previously, Maze levels were assuming that any kind of non-optimal
result implied a non-passing result; this is because we were formerly
not using the `APP_SPECIFIC_ACCEPTABLE_FAIL` result for any Maze levels.

I thought I had caught all the places where Maze was making this
assumption in https://github.com/code-dot-org/code-dot-org/pull/17428,
but it turns out I was missing one in the milestone report, the results
of which are used by the "continue" button in the success dialog to
determine whether or not clicking it will advance us on to the next
level.