Pledge map: disable cron job while geocoding isn't running
We don't know why, but when we add new entries to the pledge Fusion Table via cron job, the location entries aren't being geocoded.  Instead, they are marked in the Fusion Table UI with a yellow background, which can indicate that a quota has been used up, even though we can't find evidence of that elsewhere.

Since we reupload all entries to the Fusion Table fresh every day (overwriting existing entries, same as we do for the Hour of Code map) we lose all the dots on the map.

Until we have a better idea how to address this, we're disabling this cron job.  Instead, we will run the script manually, and trigger geocoding in the Fusion Table UI manually at that time.