Courses: smaller & better left/right margins when responsive
Since Eyes and Saucelabs aren't giving us a 1024 viewport (it's more like 1014) this attempts a more aggressive threshold for initial responsiveness.  The page body remains 970px wide, but we now allow this to be shown when the browser is 992px and higher (same as Bootstrap 3's grid).  When the browser is smaller, we now go down to 10px left/right margin.  By using a different approach to the "full width" container's styling, we center the content better too, by not having the container's CSS putting in 25px padding.