Update to bramble 0.1.21
Bump bramble by a patch version (0.1.20 to 0.1.21) picking up a small fix that allows navigation between pages to succeed when the links include anchors or queryparams.  Unfortuantely those components are still ignored, it's a known issue we'll look at in more depth down the road.

New version is published and uploaded to S3, this commit updates the dependency and download link in the main repo.  Tested manually on local machine (with npm-published bramble) and verified that workaround is functioning as expected.