Tutorial explorer: report tutorial start by grade range
As well as reporting a tutorial event action "start" with event label of tutorial code, report a tutorial event action "start-[grade]" with event label of tutorial code.

This will give us five new buckets of tutorial starts (with [grade] being "all", "pre", "2-5", "6-8" or "9+") so that we can see which tutorials are most popular in each grade range.