Don't leave RedisTable::NotFound uncaught
The NetSimApi is correctly responding with a 404 whenever a row can't be found in the Redis table, and the clients expect this and handle it gracefully.  The internal mechanism for this, though, is to raise an uncaught RedisTable::NotFound < Sinatra::NotFound exception, which results in the 404 response.

These uncaught exceptions are getting logged to Honeybadger, when we really think they're our servers responding correctly. (https://app.honeybadger.io/projects/3240/faults/29223700)

Here, I'm manually rescuing from the NotFound error and responding with a 404, in the hope that these will stop making it to Honeybadger.