Omit pairing info when driver account is deleted
A [Zendesk ticket][1] and associated [Honeybadger error][2] brought to our attention that we've got a server error in a very particular situation: If two users have made pair programming progress on a channel-backed level and then the driver's account is deleted, then when the navigator attempts to revisit that level a 500 will be served when we try to pull up a channel id for the deleted account.

After consulting with @poorvasingal we've decided it's okay to omit any pair-programming information in this situation, which just means the navigator doesn't get a toast offering to show them the solution they and their pair came up with in the past.  That also seems like the original intent of the code, which checks whether the returned driver is falsy but did not anticipate getting the string `"deleted user"` back.

I've added a regression test that covers this case to the dashboard unit test suite.

[1]: https://codeorg.zendesk.com/agent/tickets/115705
[2]: https://app.honeybadger.io/projects/3240/faults/35473924