Reenable MC i18n test
Previous failures were due to either the `{currentPlayerName}` special
string or standard markdown formatting being present in the instructions
string. Fortunately, the very first level has neither of those.

Unfortunately, it looks like it is the only one that does; these
instructions are supposed to remain as they are, but if either
formatting feature gets introduced in the future we're going to need to
improve this test to deal with formatted strings. For now, though, this
is good enough.