Make assets_bucket_prefix configurable
Makes it possible to customize the S3 prefix for CDN content for the current stack (which is stored in the cdo-dist bucket, and previously under a rack_env prefix).  We need to customize this because our staging-next machine (which has a 'test' rack_env) is not configured with access to any prefix besides adhoc and I don't have permission to customize that IAM role, or to export a new one.