Don't pass bad access token
Don't pass a bad and unneeded GitHub access token when trying to read what is a public resource anyway.  This URL returns the correct information if no auth token is given, but a bad credentials error if you pass an invalid token like we are right now.  I can't find this environment variable (spelled correctly or incorrectly) defined anywhere in our repository or CircleCI config so tearing it out is probably the right thing to do.