[POC] Build apps locale with English fallback
Right now, when building out apps locale files from the various locale
json, we only build out what's defined for that language and do not in
any case fall back to english definitions. This means that when we add a
new localization string to english, it won't work for any other language
until the apps/sync-apps.sh script runs, which propagates it out.

What this means is that every apps i18n change needs to have an
associated build step and git commit at a point in the build process
before we can run any tests involving i18n. For our current build setup,
this means that build and git activity happens on staging and regularly
causes merge issues and that i18n tests are all required to be disabled
on circle.

This PR is a simple proof-of-concept for a new approach, in which we
simply build out the i18n with English fallback, with the expectation
that the new strings will be added to the locale-specific json files as
part of the i18n sync process with crowdin rather than as a build step