Fix tests
The resize bar tests pass in isolation but fail when run in the context of our full suite.  This is because testing the resize bar involves firing some events on window, and in lots of other places in our tests we set up singletons like StudioApp that listen to these events and don't get cleaned up properly.

Tracking down all the places StudioApp global is used is going to be prohibitively difficult, and I do actually want the component to fire events in unit tests since it's pretty core functionality.  So for now I'm making the places breaking in response to this a little more resilient when they're called at an odd time.