Remove sync-apps script
Follow-up to https://github.com/code-dot-org/code-dot-org/pull/19445

Now that apps translation files are automatically built with english
fallback, it's no longer necessary to explicitly propagate newly-added
i18n strings out to every locale.

Note that because it used to be the case that this required sync only
happened on a staging build, circle tests were formerly unable test
i18n. Therefore, next steps are to identify tests that have been flagged
with no-circle but no longer need to be.