Header: remove CSS that was only used for legacy header
And rename header2* to header* now that we have fully switched to the new header.

This is a new attempt at the fix.  It includes a change to the Pegasus route handler for style.css so that it uses an etag header instead of a last-modified header, which was not updating because in this case we are removing a .css file from the set of files that is concatenated, rather than udpating one of them.