School autocomplete handles first server response
Fixes a bug where the school autocomplete dropdown would appear to ignore the first server response it receieved, making the initial search a no-op.

The root cause is in the use of lodash `debounce` to avoid sending a request to the server until the user has stopped typing for a moment (200ms).  The dropdown is depending on the debounced method to return a Promise that resolves to a search result, but because there's no leading-edge call on the debounced method it returns `undefined` instead of a Promise when it is initially called.

The solution is to wrap the debounced method in a new promise each time, to guarantee that a promise is returned on every call to `getOptions`.  That promise resolves when the debounced method is eventually called, or it's replaced when the user types some more, changing the search query.