Header: show sign-in/hamburger on mobile devices for responsive pages
The studio.code.org header, when on a mobile width device (defined as <= 500px wide) continues to hide the header and various items inside of it in various cases.

Except now, when on a page that is responsive (currently limited to /courses and /congrats), the right part of the header containing the sign-in button and the hamburger, are no longer hidden.