Temporarily restore weird blockly and flappy instructions
These instructions, since being moved over to dashboard rather than
apps, have been having a variety of i18n issues. After some conversation
with Crowdin support, I've determined that the problem is that these new
strings are actually marked within their system as duplicates of the
original set of strings, which are marked as 'masters'. In order to
"upgrade" this set of strings to master, I need to first reset them to
exactly match the original strings and then delete the originals,
allowing the new ones to be upgraded to master.